backend/requirements.txt
# ─────────────────────────────────────────────
#  Prox Challenge — Vulcan OmniPro 220 QA Agent
#  requirements.txt
# ─────────────────────────────────────────────

# API layer
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12   # required for file upload endpoints

# Anthropic Claude Agent SDK (the core LLM backbone)
anthropic>=0.40.0

# PDF parsing — OCR via Mistral
mistralai>=1.2.0,<2.0.0    # Mistral OCR API (mistral-ocr-latest)

# PDF parsing — text + image extraction (kept for utilities imported by mistral_ocr.py)
pymupdf>=1.24.0            # fitz: fast, accurate, image-export capable
pdfplumber>=0.11.0         # complementary: superior table extraction

# Vector store (no separate server required — embedded ChromaDB for local dev)
chromadb>=0.5.0

# OpenTelemetry — all pins must match; resolves chromadb import conflict on Python 3.13
opentelemetry-sdk==1.41.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-proto==1.41.0

# Embeddings (local, no second API key needed)
sentence-transformers>=3.3.0
torch>=2.2.0               # sentence-transformers dependency

# Knowledge graph
networkx>=3.4.0

# Schema / validation / settings
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Image handling
pillow>=11.0.0
numpy>=2.0.0

# HTTP client (for async ingestion calls)
httpx>=0.28.0

# Environment
python-dotenv>=1.0.1
